Move test classes into a quasar-core "tests" jar.
