+++ quasar-core/src/test/java/co/paralleluniverse/common/resource/BestClassLoaderTest.java
@@ -1,5 +1,6 @@
 package co.paralleluniverse.common.resource;
 
+import co.paralleluniverse.common.test.ClassFactory;
 import org.junit.Test;
 
 import java.net.URL;
