+++ quasar-core/src/test/java/co/paralleluniverse/common/test/TestInterface.java
