+++ quasar-core/src/test/java/co/paralleluniverse/common/test/TestInterface2.java
